+++ tiger-rbel/src/main/java/de/gematik/rbellogger/key/RbelKeyManager.java
@@ -7,12 +7,11 @@ package de.gematik.rbellogger.key;
 import de.gematik.rbellogger.converter.RbelConverterPlugin;
 import de.gematik.rbellogger.data.RbelElement;
 import de.gematik.rbellogger.data.facet.RbelJsonFacet;
-import lombok.extern.slf4j.Slf4j;
-
-import javax.crypto.spec.SecretKeySpec;
 import java.security.Key;
 import java.util.*;
 import java.util.stream.Stream;
+import javax.crypto.spec.SecretKeySpec;
+import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 public class RbelKeyManager {
@@ -25,9 +24,13 @@ public class RbelKeyManager {
                     .map(RbelElement::getRawStringContent)
                     .map(tokenB64 -> {
                         try {
-                            return Base64.getDecoder().decode(tokenB64);
-                        } catch (Exception e) {
-                            return null;
+                            return Base64.getUrlDecoder().decode(tokenB64);
+                        } catch (Exception e1) {
+                            try {
+                                return Base64.getDecoder().decode(tokenB64);
+                            } catch (Exception e2) {
+                                return null;
+                            }
                         }
                     })
                     .filter(Objects::nonNull)